LayoutTests/fast/js/array-reset-large-index-expected.txt

This test puts an item in a big index and then tries to change it. It shoudl change successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array[10001] is 5
PASS array[10002] is "b"
PASS successfullyParsed is true

TEST COMPLETE
